backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==1.4.50
pydantic==1.10.13
alembic==1.12.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
pytest==7.4.3
httpx==0.25.1
aiosqlite==0.19.0
python-dotenv==1.0.0
email-validator==2.1.0
requests==2.31.0
